#4092 - Move getTimeInCurrentTimezone to a separate Util class for working with dates
